Split analysis backfill trigger modules
