+++ src/app/controller/library/analysis_backfill.rs
@@ -4,6 +4,8 @@ use super::*;
 use std::collections::BTreeMap;
 use std::path::{Path, PathBuf};
 
+mod trigger;
+
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub(crate) enum AnalysisTriggerReason {
     SampleAdded,
@@ -243,319 +245,11 @@ impl AppController {
         self.spawn_analysis_trigger(AnalysisTrigger::UserRequestedReanalysis { source, action });
     }
 
-    fn spawn_analysis_trigger(&mut self, trigger: AnalysisTrigger) {
-        let tx = self.runtime.jobs.message_sender();
-        std::thread::spawn(move || match trigger {
-            AnalysisTrigger::ChangedSamples {
-                source,
-                changed_samples,
-                announce,
-            } => {
-                let changed_samples: Vec<_> = changed_samples
-                    .iter()
-                    .map(ChangedSampleInput::to_changed_sample)
-                    .collect();
-                let result = analysis_jobs::enqueue_jobs_for_source(&source, &changed_samples);
-                send_changed_sample_enqueue_result(tx, result, announce);
-            }
-            AnalysisTrigger::UserRequestedReanalysis { source, action } => match action {
-                ManualReanalysisAction::SelectedSource => {
-                    let result = analysis_jobs::enqueue_jobs_for_source_backfill_full(&source);
-                    send_changed_sample_enqueue_result(tx.clone(), result, true);
-
-                    let result = analysis_jobs::enqueue_jobs_for_embedding_backfill(&source);
-                    send_embedding_enqueue_result(tx, result, true);
-                }
-                ManualReanalysisAction::SelectedRows {
-                    changed_samples,
-                    sample_ids,
-                } => {
-                    if !changed_samples.is_empty() {
-                        let changed_samples: Vec<_> = changed_samples
-                            .iter()
-                            .map(ChangedSampleInput::to_changed_sample)
-                            .collect();
-                        let result =
-                            analysis_jobs::enqueue_jobs_for_source(&source, &changed_samples);
-                        send_changed_sample_enqueue_result(tx.clone(), result, true);
-                    }
-                    let result =
-                        analysis_jobs::enqueue_jobs_for_embedding_samples(&source, &sample_ids);
-                    send_embedding_enqueue_result(tx, result, true);
-                }
-                ManualReanalysisAction::SimilarityPrepBootstrap {
-                    force_full_analysis,
-                } => {
-                    let analysis_result = if force_full_analysis {
-                        analysis_jobs::enqueue_jobs_for_source_backfill_full(&source)
-                    } else {
-                        analysis_jobs::enqueue_jobs_for_source_backfill(&source)
-                    };
-                    send_changed_sample_enqueue_result(tx.clone(), analysis_result, true);
-
-                    let embed_result = analysis_jobs::enqueue_jobs_for_embedding_backfill(&source);
-                    send_embedding_enqueue_result(tx, embed_result, true);
-                }
-            },
-        });
-    }
-
     /// Return true if any sources are configured.
     pub fn has_any_sources(&self) -> bool {
         !self.library.sources.is_empty()
     }
 }
 
-fn send_changed_sample_enqueue_result(
-    tx: super::jobs::JobMessageSender,
-    result: Result<(usize, analysis_jobs::AnalysisProgress), String>,
-    announce: bool,
-) {
-    match result {
-        Ok((inserted, progress)) => {
-            let _ = tx.send(super::jobs::JobMessage::Analysis(
-                analysis_jobs::AnalysisJobMessage::EnqueueFinished {
-                    inserted,
-                    progress,
-                    announce,
-                },
-            ));
-        }
-        Err(err) => {
-            let _ = tx.send(super::jobs::JobMessage::Analysis(
-                analysis_jobs::AnalysisJobMessage::EnqueueFailed(err),
-            ));
-        }
-    }
-}
-
-fn send_embedding_enqueue_result(
-    tx: super::jobs::JobMessageSender,
-    result: Result<(usize, analysis_jobs::AnalysisProgress), String>,
-    announce: bool,
-) {
-    match result {
-        Ok((inserted, progress)) => {
-            let _ = tx.send(super::jobs::JobMessage::Analysis(
-                analysis_jobs::AnalysisJobMessage::EmbeddingBackfillEnqueueFinished {
-                    inserted,
-                    progress,
-                    announce,
-                },
-            ));
-        }
-        Err(err) => {
-            let _ = tx.send(super::jobs::JobMessage::Analysis(
-                analysis_jobs::AnalysisJobMessage::EmbeddingBackfillEnqueueFailed(err),
-            ));
-        }
-    }
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::app::controller::jobs::JobMessage;
-    use crate::app::controller::test_support::{dummy_controller, sample_entry, write_test_wav};
-    use std::time::{Duration, Instant};
-
-    fn wait_for_analysis_message(
-        controller: &mut AppController,
-        mut predicate: impl FnMut(&analysis_jobs::AnalysisJobMessage) -> bool,
-    ) -> analysis_jobs::AnalysisJobMessage {
-        let deadline = Instant::now() + Duration::from_secs(5);
-        while Instant::now() < deadline {
-            match controller.runtime.jobs.try_recv_message() {
-                Ok(JobMessage::Analysis(message)) if predicate(&message) => return message,
-                Ok(_) => {}
-                Err(std::sync::mpsc::TryRecvError::Empty) => {
-                    std::thread::sleep(Duration::from_millis(10));
-                }
-                Err(err) => panic!("unexpected receive error: {err:?}"),
-            }
-        }
-        panic!("timed out waiting for analysis message");
-    }
-
-    fn pending_job_count(source: &SampleSource, job_type: &str) -> i64 {
-        analysis_jobs::db::open_source_db(&source.root)
-            .unwrap()
-            .query_row(
-                "SELECT COUNT(*) FROM analysis_jobs WHERE job_type = ?1 AND status = 'pending'",
-                rusqlite::params![job_type],
-                |row| row.get(0),
-            )
-            .unwrap()
-    }
-
-    fn prepare_manual_reanalysis_fixture(
-        entries: &[&str],
-    ) -> (
-        AppController,
-        SampleSource,
-        tempfile::TempDir,
-        crate::app_dirs::ConfigBaseGuard,
-    ) {
-        let config_dir = tempfile::tempdir().unwrap();
-        let guard = crate::app_dirs::ConfigBaseGuard::set(config_dir.path().to_path_buf());
-        let (mut controller, source) = dummy_controller();
-        controller.library.sources.push(source.clone());
-        let wav_entries: Vec<_> = entries
-            .iter()
-            .map(|entry| {
-                let path = source.root.join(entry);
-                if let Some(parent) = path.parent() {
-                    std::fs::create_dir_all(parent).unwrap();
-                }
-                write_test_wav(&path, &[0.0, 1.0, 0.0, -1.0]);
-                sample_entry(entry, crate::sample_sources::Rating::NEUTRAL)
-            })
-            .collect();
-        controller.set_wav_entries_for_tests(wav_entries);
-        controller.rebuild_wav_lookup();
-        controller.rebuild_browser_lists();
-        (controller, source, config_dir, guard)
-    }
-
-    #[test]
-    fn trigger_policy_matrix_keeps_implicit_reasons_disabled() {
-        let cases = [
-            (
-                AnalysisTriggerReason::SampleAdded,
-                AnalysisTriggerPolicy::ChangedSamples,
-            ),
-            (
-                AnalysisTriggerReason::AudioContentChanged,
-                AnalysisTriggerPolicy::ChangedSamples,
-            ),
-            (
-                AnalysisTriggerReason::UserRequestedReanalysis,
-                AnalysisTriggerPolicy::UserRequestedReanalysis,
-            ),
-            (
-                AnalysisTriggerReason::SimilarityPrepBootstrap,
-                AnalysisTriggerPolicy::SimilarityPrepBootstrap,
-            ),
-            (
-                AnalysisTriggerReason::ScanCompleted,
-                AnalysisTriggerPolicy::Forbidden,
-            ),
-            (
-                AnalysisTriggerReason::WatcherAutoSync,
-                AnalysisTriggerPolicy::Forbidden,
-            ),
-            (
-                AnalysisTriggerReason::DeferredMaintenance,
-                AnalysisTriggerPolicy::Forbidden,
-            ),
-            (
-                AnalysisTriggerReason::RenameWithoutContentChange,
-                AnalysisTriggerPolicy::Forbidden,
-            ),
-            (
-                AnalysisTriggerReason::SimilarityReadPath,
-                AnalysisTriggerPolicy::Forbidden,
-            ),
-        ];
-
-        for (reason, expected) in cases {
-            assert_eq!(reason.policy(), expected, "reason={reason:?}");
-        }
-    }
-
-    #[test]
-    fn manual_selected_source_reanalysis_enqueues_analysis_and_embeddings() {
-        let (mut controller, source, _config_dir, _guard) =
-            prepare_manual_reanalysis_fixture(&["Pack/a.wav", "Pack/b.wav"]);
-
-        controller.reanalyze_selected_source();
-
-        match wait_for_analysis_message(&mut controller, |message| {
-            matches!(
-                message,
-                analysis_jobs::AnalysisJobMessage::EnqueueFinished { .. }
-            )
-        }) {
-            analysis_jobs::AnalysisJobMessage::EnqueueFinished {
-                inserted, announce, ..
-            } => {
-                assert_eq!(inserted, 2);
-                assert!(announce);
-            }
-            other => panic!("unexpected analysis message: {other:?}"),
-        }
-        match wait_for_analysis_message(&mut controller, |message| {
-            matches!(
-                message,
-                analysis_jobs::AnalysisJobMessage::EmbeddingBackfillEnqueueFinished { .. }
-            )
-        }) {
-            analysis_jobs::AnalysisJobMessage::EmbeddingBackfillEnqueueFinished {
-                inserted,
-                announce,
-                ..
-            } => {
-                assert_eq!(inserted, 1);
-                assert!(announce);
-            }
-            other => panic!("unexpected embedding message: {other:?}"),
-        }
-        assert_eq!(
-            pending_job_count(&source, analysis_jobs::db::ANALYZE_SAMPLE_JOB_TYPE),
-            2
-        );
-        assert_eq!(
-            pending_job_count(&source, analysis_jobs::db::EMBEDDING_BACKFILL_JOB_TYPE),
-            1
-        );
-    }
-
-    #[test]
-    fn manual_row_reanalysis_enqueues_only_selected_visible_rows() {
-        let (mut controller, source, _config_dir, _guard) =
-            prepare_manual_reanalysis_fixture(&["Pack/a.wav", "Pack/b.wav", "Pack/c.wav"]);
-
-        controller
-            .reanalyze_browser_rows(&[0, 2, 2, usize::MAX])
-            .unwrap();
-
-        match wait_for_analysis_message(&mut controller, |message| {
-            matches!(
-                message,
-                analysis_jobs::AnalysisJobMessage::EnqueueFinished { .. }
-            )
-        }) {
-            analysis_jobs::AnalysisJobMessage::EnqueueFinished {
-                inserted, announce, ..
-            } => {
-                assert_eq!(inserted, 2);
-                assert!(announce);
-            }
-            other => panic!("unexpected analysis message: {other:?}"),
-        }
-        match wait_for_analysis_message(&mut controller, |message| {
-            matches!(
-                message,
-                analysis_jobs::AnalysisJobMessage::EmbeddingBackfillEnqueueFinished { .. }
-            )
-        }) {
-            analysis_jobs::AnalysisJobMessage::EmbeddingBackfillEnqueueFinished {
-                inserted,
-                announce,
-                ..
-            } => {
-                assert_eq!(inserted, 1);
-                assert!(announce);
-            }
-            other => panic!("unexpected embedding message: {other:?}"),
-        }
-        assert_eq!(
-            pending_job_count(&source, analysis_jobs::db::ANALYZE_SAMPLE_JOB_TYPE),
-            2
-        );
-        assert_eq!(
-            pending_job_count(&source, analysis_jobs::db::EMBEDDING_BACKFILL_JOB_TYPE),
-            1
-        );
-    }
-}
+mod tests;
